functions/logger/requirements.txt
cachecontrol==0.12.10; python_version >= "3.6"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
certifi==2021.10.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
firebase-admin==5.2.0; python_version >= "3.6"
google-api-core==2.5.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-api-python-client==2.37.0; python_version >= "3.6"
google-auth-httplib2==0.1.0; python_version >= "3.6"
google-auth==2.6.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
google-cloud-core==2.2.2; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-cloud-firestore==2.3.4; platform_python_implementation != "PyPy" and python_version >= "3.6"
google-cloud-storage==2.1.0; python_version >= "3.7"
google-cloud==0.34.0
google-crc32c==1.3.0; python_version >= "3.7"
google-resumable-media==2.2.1; python_version >= "3.7"
googleapis-common-protos==1.54.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
grpcio-status==1.43.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
grpcio==1.43.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
idna==3.3; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
msgpack==1.0.3; python_version >= "3.6"
packaging==21.3; platform_python_implementation != "PyPy" and python_version >= "3.6"
proto-plus==1.20.0; platform_python_implementation != "PyPy" and python_version >= "3.6"
protobuf==3.19.4; platform_python_implementation != "PyPy" and python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy") or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
pyparsing==3.0.7; python_version >= "3.6" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.4.0" and python_version >= "3.6" and platform_python_implementation != "PyPy"
requests==2.27.1; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy")
sanitize-filename==1.2.0; python_version >= "3.7" and python_version < "4.0"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" and platform_python_implementation != "PyPy"
